+++ src/methods/$consume/$consume.d.ts
@@ -1,13 +1,6 @@
 import { $Promise } from '../../../generate/async.macro';
 import { $SourceIterable } from '../../types/$iterable';
 
-declare function $consume<T>(
-  func: (item: T, i: number) => void,
-): (iterable: $SourceIterable<T>) => $Promise<void>;
-
-declare function $consume<T>(
-  func: (item: T, i: number) => void,
-  iterable: $SourceIterable<T>,
-): $Promise<void>;
+declare function $consume<T>(iterable: $SourceIterable<T>): $Promise<void>;
 
 export default $consume;