+++ src/methods/$consume/__tests__/consume.auto.spec.ts
@@ -10,24 +10,15 @@ import { consume } from '../../..';
 describe('consume', () => {
   it('consumes an iterable', () => {
     const arr: Array<number> = [];
-    consume(item => arr.push(item), [1, 2, 3]);
-    expect(arr).toEqual([1, 2, 3]);
-  });
-  it('consumes an iterable using a promise', () => {
-    const arr: Array<number> = [];
     consume(
-      item => {
-        arr.push(item);
-        return Promise.resolve(0);
-      },
-      [1, 2, 3],
+      (function*() {
+        arr.push(1);
+        yield;
+        arr.push(2);
+        yield;
+        arr.push(3);
+      })(),
     );
-    expect(arr).toEqual([1, 2, 3]);
-  });
-  it('consumes an iterable (curried)', () => {
-    const arr: Array<number> = [];
-    const consumePush = consume((item: number) => arr.push(item));
-    consumePush([1, 2, 3]);
     expect(arr).toEqual([1, 2, 3]);
   });
 });