+++ src/methods/$consume/async-consume.d.ts
@@ -7,11 +7,5 @@
  */
 
 import { AsyncSourceIterable } from '../../types/async-iterable';
-declare function asyncConsume<T>(
-  func: (item: T, i: number) => void,
-): (iterable: AsyncSourceIterable<T>) => Promise<void>;
-declare function asyncConsume<T>(
-  func: (item: T, i: number) => void,
-  iterable: AsyncSourceIterable<T>,
-): Promise<void>;
+declare function asyncConsume<T>(iterable: AsyncSourceIterable<T>): Promise<void>;
 export default asyncConsume;