+++ src/methods/$for-each/$for-each.d.ts
@@ -1,0 +1,13 @@
+import { $Promise } from '../../../generate/async.macro';
+import { $SourceIterable } from '../../types/$iterable';
+
+declare function $forEach<T>(
+  callback: (item: T, i: number) => void,
+): (iterable: $SourceIterable<T>) => $Promise<void>;
+
+declare function $forEach<T>(
+  callback: (item: T, i: number) => void,
+  iterable: $SourceIterable<T>,
+): $Promise<void>;
+
+export default $forEach;